Extract DynamoUtil from DynamoDocker with createTable and removeTable
Move createTable logic into a standalone DynamoUtil class so table
management is not tied to the Docker test helper. Add removeTable for
cleanup. DynamoDocker delegates to DynamoUtil for backwards compat.
